backend/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
sqlalchemy>=2.0.40
psycopg[binary]==3.2.10
pydantic>=2.11
pydantic-settings>=2.7
python-jose[cryptography]==3.3.0
bcrypt>=4.2.0
python-multipart==0.0.19
email-validator==2.2.0
